Added produxt and reseller ordering in account extension. Deleted unnecessary comments
